Cumulative update 7 Mar 22
Cumulative update. Includes Untitled notebook which converts a star's right ascension, declination, and distance into x, y, and z coordinates for graphing. This is similar to the pdf that was provided with the project from Codecademy, but this can now be extended to any star with the correct data to input.